+++ script.node/cmd.all.clean.ts
@@ -1,4 +1,4 @@
-import { Builder, pc, Util } from './common/index.mjs';
+import { Builder, pc, Util } from './common';
 
 /**
  * Run
