+++ script.node/cmd.all.list.ts
@@ -1,4 +1,4 @@
-import { Builder, LogTable, Util, fs, minimist, pc } from './common/index.mjs';
+import { Builder, LogTable, Util, fs, minimist, pc } from './common';
 
 const argv = minimist(process.argv.slice(2));
