+++ script.node/cmd.test.ts
@@ -1,4 +1,4 @@
-import { Builder, minimist } from './common/index.mjs';
+import { Builder, minimist } from './common';
 
 const dir = process.cwd();
 const argv = minimist(process.argv.slice(2));
