--- script.node/common/index.mts
@@ -1,5 +1,0 @@
-import type * as t from './types.mjs';
-
-export * from './libs.mjs';
-export { Util } from './util.mjs';
-export { type t };
