+++ script.node/common/index.ts
@@ -1,0 +1,4 @@
+export type * as t from './t';
+
+export * from './libs';
+export { Util } from './u';
